Fix: Response time durations were incorrect after 71 minutes of execution due to a 32-bit overflow.
git-svn-id: svn://svn.code.sf.net/p/sipp/code/sipp/trunk@286 60913f6b-29f6-41b8-8391-1ad2d2d6c344

+++ call.hpp
@@ -159,7 +159,7 @@ public:
   unsigned int   paused_until;
 
   unsigned long  start_time;
-  unsigned long  start_time_rtd[MAX_RTD_INFO_LENGTH];
+  unsigned long  long start_time_rtd[MAX_RTD_INFO_LENGTH];
 
   bool           rtd_done[MAX_RTD_INFO_LENGTH];
